rename DailyDocketRow
